fix: update product and related modules

+++ src/modules/customer/customer-address/customer-address.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateCustomerAddressDto } from './dto/create-customer-address.dto';
 import { UpdateCustomerAddressDto } from './dto/update-customer-address.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -18,13 +18,46 @@ export class CustomerAddressService {
     private readonly locationRegionsService: LocationRegionsService,
   ) {}
 
-  async create(dto: CreateCustomerAddressDto) {
-    const { customer, city, country, district, ward } = dto;
+  async findAll(query: CustomerAddressQueryDto) {
+    const { page, limit } = query;
 
     //
-    const [customerExist, locationExists] = await Promise.all([
-      this.customerService.exists([customer]),
-      this.locationRegionsService.exists([city, country, district, ward]),
+    const { take, skip } = calculatePagination(page, limit);
+
+    const [data, totalData] = await this.customerAddressRepo.findAndCount({
+      relations: ['customer', 'country', 'provinceCity', 'districtTown', 'wardCommune'],
+      select: {
+        id: true,
+        address: true,
+        createdAt: true,
+        isDefault: true,
+        customer: { id: true, fullname: true, phone: true },
+        country: { id: true, name: true },
+        provinceCity: { id: true, name: true },
+        districtTown: { id: true, name: true },
+        wardCommune: { id: true, name: true },
+      },
+      order: { createdAt: 'DESC' },
+      take,
+      skip,
+    });
+
+    return {
+      data,
+      totalData,
+      page,
+      totalPage: Math.ceil(totalData / limit),
+    };
+  }
+
+  async create(id: string, dto: CreateCustomerAddressDto) {
+    const { provinceCity, country, districtTown, wardCommune, ...rest } = dto;
+
+    //
+    const [customerExist, isFirstAddress, locationExists] = await Promise.all([
+      this.customerService.exists([id]),
+      this.customerAddressRepo.count(),
+      this.locationRegionsService.exists([provinceCity, country, districtTown, wardCommune]),
     ]);
 
     //
@@ -35,46 +68,16 @@ export class CustomerAddressService {
 
     //
     const dataCreate = this.customerAddressRepo.create({
-      customer: { id: customer },
-      country: { id: country },
-      city: { id: city },
-      district: { id: district },
-      ward: { id: ward },
+      ...rest,
+      customer: { id: id },
       address: dto.address,
+      country: { id: country },
+      wardCommune: { id: wardCommune },
+      provinceCity: { id: provinceCity },
+      districtTown: { id: districtTown },
+      isDefault: isFirstAddress === 0 ? true : dto.isDefault,
     });
     return this.customerAddressRepo.save(dataCreate);
-  }
-
-  async findAll(query: CustomerAddressQueryDto) {
-    const { page, limit } = query;
-
-    //
-    const { take, skip } = calculatePagination(page, limit);
-
-    const [data, totalData] = await this.customerAddressRepo.findAndCount({
-      relations: ['customer', 'country', 'city', 'district', 'ward'],
-      select: {
-        id: true,
-        address: true,
-        createdAt: true,
-        isDefault: true,
-        customer: { id: true, fullname: true, phone: true },
-        country: { id: true, name: true },
-        city: { id: true, name: true },
-        district: { id: true, name: true },
-        ward: { id: true, name: true },
-      },
-      order: { createdAt: 'DESC' },
-      take,
-      skip,
-    });
-
-    return {
-      data,
-      totalData,
-      page,
-      totalPage: Math.ceil(totalData / limit),
-    };
   }
 
   async findAllOwned(customerId: string, query: CustomerAddressQueryDto) {
@@ -85,17 +88,19 @@ export class CustomerAddressService {
 
     const [data, totalData] = await this.customerAddressRepo.findAndCount({
       where: { customer: { id: customerId } },
-      relations: ['customer', 'country', 'city', 'district', 'ward'],
+      relations: ['customer', 'country', 'provinceCity', 'districtTown', 'wardCommune'],
       select: {
         id: true,
         address: true,
         createdAt: true,
         isDefault: true,
+        recipientName: true,
+        recipientPhone: true,
         customer: { id: true, fullname: true, phone: true },
         country: { id: true, name: true },
-        city: { id: true, name: true },
-        district: { id: true, name: true },
-        ward: { id: true, name: true },
+        provinceCity: { id: true, name: true },
+        districtTown: { id: true, name: true },
+        wardCommune: { id: true, name: true },
       },
       order: { createdAt: 'DESC' },
       take,
@@ -110,33 +115,45 @@ export class CustomerAddressService {
     };
   }
 
-  async update(id: string, dto: UpdateCustomerAddressDto) {
-    const { customer, city, country, district, ward } = dto;
+  async updateOwned(id: string, customerId: string, dto: UpdateCustomerAddressDto) {
+    const { provinceCity, country, districtTown, wardCommune, isDefault, ...restDto } = dto;
 
     //
-    const [customerExist, ...rest] = await Promise.all([
-      customer ? this.customerService.exists([customer]) : null,
-      ward ? this.locationRegionsService.exists([ward]) : null,
-      city ? this.locationRegionsService.exists([city]) : null,
+    const [customerExist, existWard, existProvince, existCountry, existDistrict] = await Promise.all([
+      this.customerService.exists([customerId]),
+      wardCommune ? this.locationRegionsService.exists([wardCommune]) : null,
+      provinceCity ? this.locationRegionsService.exists([provinceCity]) : null,
       country ? this.locationRegionsService.exists([country]) : null,
-      district ? this.locationRegionsService.exists([district]) : null,
+      districtTown ? this.locationRegionsService.exists([districtTown]) : null,
     ]);
 
     //
     if (!customerExist) throw new NotFoundException('Customer not found');
-    if (!rest.every((exists) => exists)) {
+    if (
+      (wardCommune && !existWard) ||
+      (provinceCity && !existProvince) ||
+      (country && !existCountry) ||
+      (districtTown && !existDistrict)
+    ) {
       throw new NotFoundException('One or more location regions do not exist');
+    }
+
+    //
+    if (isDefault) {
+      // Nếu isDefault = true, thì set tất cả các địa chỉ khác của customer này về isDefault = false
+      await this.customerAddressRepo.update({ customer: { id: customerId }, isDefault: true }, { isDefault: false });
     }
 
     //
     const dataUpdate = await this.customerAddressRepo.preload({
       id,
-      customer: customer ? { id: customer } : undefined,
+      ...restDto,
+      isDefault,
+      customer: { id: customerId },
       country: country ? { id: country } : undefined,
-      city: city ? { id: city } : undefined,
-      district: district ? { id: district } : undefined,
-      ward: ward ? { id: ward } : undefined,
-      address: dto.address,
+      provinceCity: provinceCity ? { id: provinceCity } : undefined,
+      districtTown: districtTown ? { id: districtTown } : undefined,
+      wardCommune: wardCommune ? { id: wardCommune } : undefined,
     });
 
     if (!dataUpdate) {
@@ -146,11 +163,17 @@ export class CustomerAddressService {
     return await this.customerAddressRepo.save(dataUpdate);
   }
 
-  async remove(id: string) {
-    const customerAddress = await this.customerAddressRepo.findOne({ where: { id } });
+  async removeOwned(id: string, customerId: string) {
+    const customerAddress = await this.customerAddressRepo.findOne({ where: { id, customer: { id: customerId } } });
+
     if (!customerAddress) {
       throw new NotFoundException('CustomerAddress not found');
     }
+
+    if (customerAddress.isDefault) {
+      throw new BadRequestException('Cannot delete default address');
+    }
+
     return await this.customerAddressRepo.remove(customerAddress);
   }
 }
